refactor(sourcing-tests): move sourcing tests to pop-common
