+++ crates/pop-common/Cargo.toml
@@ -27,4 +27,5 @@ url.workspace = true
 
 [dev-dependencies]
 mockito.workspace = true
+strum_macros.workspace = true
 tempfile.workspace = true
